requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
Booktype==1.5
cachetools==6.2.4
certifi==2026.1.4
colorama==0.4.6
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
magic-filter==1.0.12
Menu==3.2.2
multidict==6.7.0
propcache==0.4.1
psutil==7.2.1
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
raito==1.3.7
redis>=5.0.0
simplejson==3.20.2
SQLAlchemy==2.0.45
sqlparse==0.5.5
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
watchfiles==1.1.1
yarl==1.22.0